+++ CHANGELOG.md
@@ -1,8 +1,11 @@
+# 0.2.1
+- feat: Added WidTapToHideKeyboard widget 🎉
+- feat: Implemented AppBarThemes, BottomAppBarThemes, CheckboxThemes, PopupMenuThemes🎉
+- feat: Added widgets to example view 🎉
 # 0.2.0
 
 - feat: Added custom button with 4 variants 🎉
 - feat: Added several theme implementations for several widgets 🎉
-- feat: Added WidTapToHideKeyboard widget 🎉
 - feat: Added custom touchable widget 🎉
 - feat: Exposing and providing fonts 🎉